api/requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
Bootstrap-Flask==1.5.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
entrypoints==0.3
Flask==0.12.5
flask-appconfig==0.11.1
Flask-Bootstrap==3.3.7.1
Flask-Debug==0.4.3
flask-nav==0.6
Flask-WTF==0.14.3
future==0.18.2
gast==0.3.3
grpcio==1.33.2
gunicorn
h5py==2.10.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.1.0
importlib-resources==3.3.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
rdflib==5.0.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.25.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.1
tqdm==4.53.0
traitlets==4.3.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.0
flask_swagger_ui
